Import control-center 2.8.2.
git-svn-id: svn+ssh://svn.mit.edu/athena/branches/vendor@21318 728af825-273c-0410-89f9-f7d3b574a069

+++ third/control-center/libgswitchit/gswitchit_util.c
@@ -60,8 +60,8 @@ GSwitchItLogAppender (const char file[], const char function[],
 		      int level, const char format[], va_list args)
 {
 	time_t now = time (NULL);
-	g_log (NULL, G_LOG_LEVEL_DEBUG, "[%08ld,%03d,%s:%s/] \t", now,
-	       level, file, function);
+	g_log (NULL, G_LOG_LEVEL_DEBUG, "[%08ld,%03d,%s:%s/] \t",
+	       (long)now, level, file, function);
 	g_logv (NULL, G_LOG_LEVEL_DEBUG, format, args);
 }
